pyproject.toml
[tool.poetry]
name = "pdf-splitter-agent"
version = "0.1.0"
description = "An AI-powered agent to intelligently split multi-page PDF documents using LangGraph and LLMs."
authors = ["Your Name <you@example.com>"]
readme = "README.md"
license = "MIT"
package-mode = false

[tool.poetry.dependencies]
python = "^3.9"

# Core LangChain and LangGraph libraries
langchain = "^0.3.27"
langgraph = "^0.6.5"
langchain-core = "^0.3.74"

# LLM and Embedding integrations
langchain-community = "^0.3.27"
langchain-google-genai = "^2.1.9"
langchain-ollama = "^0.3.6"

# PDF and Environment management
pypdf2 = "^3.0.1"
python-dotenv = "^1.1.1"

# Database and Vector Store
pymongo = "^4.14.0"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"